requirements.txt
aws-cdk-lib==2.68.0
constructs>=10.0.0,<11.0.0
flake8==6.0.0
black==23.1.0
